+++ test/test_render.jl
@@ -912,5 +912,17 @@ end
         @test u[1] == u.tree.children[1]
         @test u[1, 1] == u.tree.children[1].children[1]
         @test u[1, 1, 1] == u.tree.children[1].children[1].children[1]
+        # Footprint uses outer contour when there's only one
+        dr1 = difference2d(r2, r4)
+        @test Polygons.circularapprox(
+            rotate(footprint(dr1), 45°).p,
+            rotate(r2, 45°).p,
+            atol=1e-9μm
+        )
+        # Halo uses original ClippedPolygon, hole in the center
+        # Offset returns holes as reversed-orientation polygons [issue #11]
+        @test Polygons.orientation(halo(dr1, 0.1μm)[2]) == -1 # still has a hole
+        @test footprint(union2d(r1, r1 + Point(40, 0)μm)) isa Rectangle # multipolygon => use bounds
+        @test halo(union2d(r3), 1μm, -0.5μm) == dr1 # ClippedPolygon halo with inner delta
     end
 end